+++ packages/npm/oa42-generator/src/generators/files/client-ts.ts
@@ -1,9 +1,9 @@
 import { banner } from "@oa42/core";
 import { RouterMode } from "goodrouter";
 import * as models from "../../models/index.js";
-import { packageInfo, toCamel, toPascal } from "../../utils/index.js";
+import { packageInfo } from "../../utils/index.js";
 import { itt } from "../../utils/iterable-text-template.js";
-import { generateClientOperationFunctionBody } from "../functions/index.js";
+import { generateClientOperationFunction } from "../functions/client-operation.js";
 import {
   generateOperationCredentialsType,
   generateOperationIncomingResponseType,
@@ -49,32 +49,7 @@ export function* generateClientTsCode(apiModel: models.Api) {
 
   for (const pathModel of apiModel.paths) {
     for (const operationModel of pathModel.operations) {
-      const operationFunctionName = toCamel(operationModel.name);
-      const operationOutgoingRequestName = toPascal(operationModel.name, "outgoing", "request");
-      const operationIncomingResponseName = toPascal(operationModel.name, "incoming", "response");
-      const credentialsName = toPascal(operationModel.name, "credentials");
-
-      const jsDoc = [
-        operationModel.deprecated ? "@deprecated" : "",
-        operationModel.summary,
-        operationModel.description,
-      ]
-        .map((line) => line.trim())
-        .filter((line) => line.length > 0)
-        .join("\n");
-
-      yield itt`
-        /**
-          ${jsDoc}
-         */
-        export async function ${operationFunctionName}(
-          outgoingRequest: ${operationOutgoingRequestName},
-          credentials: ${credentialsName},
-          configuration: ClientConfiguration = defaultClientConfiguration,
-        ): Promise<${operationIncomingResponseName}> {
-          ${generateClientOperationFunctionBody(apiModel, pathModel, operationModel)}
-        }
-      `;
+      yield* generateClientOperationFunction(apiModel, pathModel, operationModel);
       yield* generateOperationCredentialsType(apiModel, operationModel);
       yield* generateOperationOutgoingRequestType(apiModel, operationModel);
       yield* generateOperationIncomingResponseType(apiModel, operationModel);
