+++ packages/npm/oa42-generator/src/generators/files/parameters-ts.ts
@@ -1,10 +1,10 @@
 import { banner } from "@oa42/core";
 import * as models from "../../models/index.js";
-import { packageInfo, toCamel, toPascal } from "../../utils/index.js";
+import { packageInfo } from "../../utils/index.js";
 import { itt } from "../../utils/iterable-text-template.js";
 import {
-  generateIsRequestParametersFunctionBody,
-  generateIsResponseParametersFunctionBody,
+  generateIsRequestParametersFunction,
+  generateIsResponseParametersFunction,
 } from "../functions/index.js";
 import {
   generateOperationParametersTypes,
@@ -53,64 +53,11 @@ export function* generateParametersTsCode(apiModel: models.Api) {
 
   for (const pathModel of apiModel.paths) {
     for (const operationModel of pathModel.operations) {
-      const isRequestParametersFunctionName = toCamel(
-        "is",
-        operationModel.name,
-        "request",
-        "parameters",
-      );
-
-      const parseRequestParametersFunctionName = toCamel(
-        "parse",
-        operationModel.name,
-        "request",
-        "parameters",
-      );
-
-      const requestParametersTypeName = toPascal(operationModel.name, "request", "parameters");
-
-      yield itt`
-        export function ${isRequestParametersFunctionName}(
-          parameters: Partial<Record<keyof ${requestParametersTypeName}, unknown>>,
-        ): parameters is ${requestParametersTypeName} {
-          ${generateIsRequestParametersFunctionBody(apiModel, operationModel)}
-        }
-      `;
-
+      yield* generateIsRequestParametersFunction(apiModel, operationModel);
       yield* generateOperationParametersTypes(apiModel, operationModel);
 
       for (const operationResultModel of operationModel.operationResults) {
-        const isResponseParametersFunctionName = toCamel(
-          "is",
-          operationModel.name,
-          operationResultModel.statusKind,
-          "response",
-          "parameters",
-        );
-
-        const parseResponseParametersFunctionName = toCamel(
-          "parse",
-          operationModel.name,
-          operationResultModel.statusKind,
-          "response",
-          "parameters",
-        );
-
-        const responseParametersTypeName = toPascal(
-          operationModel.name,
-          operationResultModel.statusKind,
-          "response",
-          "parameters",
-        );
-
-        yield itt`
-          export function ${isResponseParametersFunctionName}(
-            parameters: Partial<Record<keyof ${responseParametersTypeName}, unknown>>,
-          ): parameters is ${responseParametersTypeName} {
-            ${generateIsResponseParametersFunctionBody(apiModel, operationResultModel)}
-          }
-        `;
-
+        yield* generateIsResponseParametersFunction(apiModel, operationModel, operationResultModel);
         yield* generateOperationResultParameterTypes(
           apiModel,
           operationModel,
