+++ packages/npm/oa42-generator/src/generators/functions/is-authentication.ts
@@ -1,17 +1,30 @@
 import * as models from "../../models/index.js";
-import { itt, joinIterable, toCamel } from "../../utils/index.js";
+import { itt, joinIterable, toCamel, toPascal } from "../../utils/index.js";
 
-export function* generateIsAuthenticationFunctionBody(
+export function* generateIsAuthenticationFunction(
   pathModel: models.Path,
   operationModel: models.Operation,
 ) {
+  const isAuthenticationFunctionName = toCamel("is", operationModel.name, "authentication");
+  const authenticationTypeName = toPascal(operationModel.name, "authentication");
+
+  yield itt`
+    export function ${isAuthenticationFunctionName}<A extends ServerAuthentication>(
+      authentication: Partial<${authenticationTypeName}<A>>,
+    ): authentication is ${authenticationTypeName}<A> {
+      ${generateBody(pathModel, operationModel)}
+    }
+  `;
+}
+
+function* generateBody(pathModel: models.Path, operationModel: models.Operation) {
   yield itt`return ${joinIterable(generateOrRules(operationModel.authenticationRequirements), " ||\n")}`;
   return;
 }
 
 function* generateOrRules(requirements: models.AuthenticationRequirement[][]) {
   if (requirements.length === 0) {
-    yield itt`true`;
+    yield JSON.stringify(true);
   }
 
   for (const subRequirements of requirements) {
@@ -21,7 +34,7 @@ function* generateOrRules(requirements: models.AuthenticationRequirement[][]) {
 
 function* generateAndRules(subRequirements: models.AuthenticationRequirement[]) {
   if (subRequirements.length === 0) {
-    yield itt`true`;
+    yield JSON.stringify(true);
   }
 
   for (const requirement of subRequirements) {
